requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
google-api-core==2.0.1
google-auth==2.0.2
google-cloud-appengine-logging==0.1.4
google-cloud-audit-log==0.1.1
google-cloud-core==2.0.0
google-cloud-logging==2.6.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
requests==2.26.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
